tests/integration/requirements.txt
# Integration Test Dependencies
# Install with: pip install -r tests/integration/requirements.txt

# Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-timeout==2.2.0
pytest-xdist==3.5.0  # Parallel execution

# HTTP Client
httpx==0.25.2

# WebSocket Testing
websockets==12.0

# Database
sqlalchemy==2.0.35
sqlmodel==0.0.14
psycopg2-binary==2.9.9

# Async Redis
redis==5.0.1
aioredis==2.0.1

# Message Queue
aio-pika==9.4.0

# Authentication
python-jose[cryptography]==3.3.0

# JSON/Data
pydantic==2.8.0
pydantic-settings==2.4.0

# Logging
structlog==24.4.0

# Testing Utilities
faker==22.0.0  # Generate test data
freezegun==1.4.0  # Freeze time

# Performance Testing (optional)
pytest-benchmark==4.0.0
locust==2.21.0  # Load testing

# Reporting (optional)
pytest-html==4.1.1
pytest-json-report==1.5.0

# Development
python-dotenv==1.0.0
ipython==8.18.0
